Refactor job_emailer to emailer and fix import and pylint
